feature: moved mongolab connection to server.js and schema definition to usermodel.js connection to mongo lab remote db OK
