+++ package.json
@@ -18,6 +18,7 @@
   },
   "homepage": "https://github.com/DudeWheresMySemiColon/DudeWheresMySemiColon",
   "dependencies": {
+     "bcrypt-nodejs": "0.0.3",
     "angular-route": "^1.4.7",
     "body-parser": "^1.14.1",
     "express": "^4.13.3",
