+++ server/server.js
@@ -1,4 +1,6 @@
 var express = require("express");
+var mongoose = require('mongoose');
+var uriUtil = require('mongodb-uri');
 
 var app = express();
 var port = process.env.PORT || 3000;
@@ -8,3 +10,29 @@ app.listen(port);
 require('./config/middleware.js')(app, express);
 
 console.log('Server now listening on port ' + port);
+
+
+//---------------------------------------------------------------------
+//mongoLab credentials - i have these from the last project
+var dbuser = 'admin';
+var dbpassword = 'admin';
+
+//set up URI connection to mongolab
+var uristring = process.env.MONGOLAB_URI || 
+process.env.MOGOHQ_URL ||
+'mongodb://' + dbuser + ':' + dbpassword + '@ds043714.mongolab.com:43714/foodly';
+
+var mongooseUri = uriUtil.formatMongoose(uristring);
+
+var options = { server: { socketOptions: { keepAlive: 1, connectTimeoutMS: 30000 } }, 
+replset: { socketOptions: { keepAlive: 1, connectTimeoutMS : 30000 } } }; 
+
+mongoose.connect(mongooseUri, options);
+var db = mongoose.connection;
+
+db.once('open',function(){
+  console.log('connected to : ', mongooseUri);
+})
+
+
+// var Schema = mongoose.Schema;
